deletelle lisätty token tarkistus. 4.21

+++ osa4/controllers/blogs.js
@@ -40,6 +40,15 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   const id = request.params.id
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const blog = await Blog.findById(id)
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    console.log("EI ONNISTU!!")
+    return
+  }
   const result = await Blog.findByIdAndDelete(id)
 
   if (result) {
